Fixed cryptojs words to uint8array conversion in randombytes and tidy up docs
