--- lib/components/ClientState.ts
@@ -1,8 +1,0 @@
-export type ClientState = {
-    identity: string;
-    IH: Array<number>; // standard Array representation of the Uint8Array
-    A: string; // hex representation of bigint
-    a: string;
-    M1: string;
-    S: string;
-}
