--- lib/components/CompatibleCrypto.ts
@@ -1,6 +1,0 @@
-import {HashFunction} from "./HashFunction";
-
-export interface CompatibleCrypto {
-    hashFunctions: { [algo: string]: HashFunction };
-    randomBytes: (length: number) => Uint8Array;
-}
