--- lib/components/IVerifierAndSalt.ts
@@ -1,4 +1,0 @@
-export interface IVerifierAndSalt {
-    verifier: string;
-    salt: string;
-}
