--- lib/components/M1AndA.ts
@@ -1,4 +1,0 @@
-export interface M1AndA {
-    A: string;
-    M1: string;
-}
