--- lib/components/PrimeGroup.ts
@@ -1,4 +1,0 @@
-export interface PrimeGroup {
-    N: bigint; // the prime
-    g: bigint; // a generator of the multiplicative group Zn
-}
