--- lib/components/ServerState.ts
@@ -1,7 +1,0 @@
-export type ServerState = {
-    identity: string;
-    salt: string; // hex representation of bigint
-    verifier: string;
-    B: string;
-    b: string;
-}
