+++ lib/modules/crypto.ts
@@ -1,5 +1,5 @@
 import * as CryptoJS from "crypto-js";
-import {CompatibleCrypto} from "../components/CompatibleCrypto";
+import {CompatibleCrypto} from "../components/cryptoTypes";
 import {wordArrayToBytes} from "./transformations";
 
 export class Crypto {
@@ -9,7 +9,7 @@ export class Crypto {
     static compatibleCrypto(): CompatibleCrypto {
         if(!this.mCrypto)
             this.mCrypto = {
-                randomBytes: (length: number) => new Uint8Array(CryptoJS.lib.WordArray.random(length).words),
+                randomBytes: (length: number) => wordArrayToBytes(CryptoJS.lib.WordArray.random(length)),
                 hashFunctions: {
                     SHA1: (data: Uint8Array) => {
                         let result: any = CryptoJS.SHA1(CryptoJS.lib.WordArray.create(data as any));
