+++ lib/modules/routines.ts
@@ -1,4 +1,4 @@
-import {Parameters} from "./Parameters";
+import {Parameters} from "./parameters";
 import {
     generateRandomBigint,
     hash,
@@ -70,7 +70,7 @@ export class Routines {
     }
 
     /**
-     * Computes X for step 2.
+     * Computes X that will be used in step 2 from client.
      * @param salt
      * @param identityHash
      */
@@ -84,7 +84,7 @@ export class Routines {
     }
 
     /**
-     * Generates an identity based on user's identity and password.
+     * Generates a hash based on user's identity and password.
      * @param identity
      * @param password
      */
@@ -101,7 +101,7 @@ export class Routines {
     }
 
     /**
-     * Generates private key. It will be used by the server (b) or the client (a).
+     * Generates private key ("a" or "b") for the client or server.
      */
     generatePrivateValue(): bigint {
         const numBits = Math.max(256, this.parameters.NBits);
@@ -116,18 +116,18 @@ export class Routines {
     }
 
     /**
-     * Generates the public value for the client.
-     * @param a The client's private value.
+     * Generates client's public key "A".
+     * @param a The client's private key "a".
      */
     computeClientPublicValue(a: bigint): bigint {
         return modPow(this.parameters.primeGroup.g, a, this.parameters.primeGroup.N);
     }
 
     /**
-     * Generates the public value for the client.
+     * Generates the server's public key "B".
      * @param k The k.
      * @param verifier
-     * @param b The server's private value.
+     * @param b The server's private key "b".
      */
     computeServerPublicValue(k: bigint, verifier: bigint, b: bigint): bigint {
         return (
@@ -137,7 +137,7 @@ export class Routines {
     }
 
     /**
-     * Checks if public value is valid.
+     * Checks if public key is valid.
      * @param value
      */
     isValidPublicValue(value: bigint): boolean {
@@ -146,8 +146,8 @@ export class Routines {
 
     /**
      * Computes U.
-     * @param A The public value of client.
-     * @param B The public value of server.
+     * @param A The client's public key "A".
+     * @param B The server's public key "B".
      */
     computeU(A: bigint, B: bigint): bigint {
         return bytesToBigint(
@@ -159,12 +159,12 @@ export class Routines {
     }
 
     /**
-     * Computes M1 which is the client's evidence.
+     * Computes client's evidence message "M1".
      * @param identity The user's identity.
      * @param salt The random salt
-     * @param A The client's public value.
-     * @param B The server's public value.
-     * @param sessionKey The session key.
+     * @param A The client's public key "A".
+     * @param B The server's public key "B".
+     * @param sessionKey The session key "S".
      */
     computeClientEvidence(identity: string, salt: bigint, A: bigint, B: bigint, sessionKey: bigint): bigint {
         return bytesToBigint(
@@ -179,10 +179,10 @@ export class Routines {
     }
 
     /**
-     * Computes M2 which is the server's evidence.
-     * @param A The client's public value.
-     * @param M1 The client's evidence.
-     * @param sessionKey The session key.
+     * Computes server's evidence message "M2".
+     * @param A The client's public value "A".
+     * @param M1 The client's evidence message "M2".
+     * @param sessionKey The session key "S".
      */
     computeServerEvidence(A: bigint, M1: bigint, sessionKey: bigint): bigint {
         return bytesToBigint(
@@ -194,12 +194,12 @@ export class Routines {
     }
 
     /**
-     * Computes the session key S for the client.
+     * Computes the session key "S" for the client.
      * @param k The k.
      * @param x The x.
      * @param u The u.
-     * @param a The client's private value.
-     * @param B The server's public value.
+     * @param a The client's private key "a".
+     * @param B The server's public key "B".
      */
     computeClientSessionKey(k: bigint, x: bigint, u: bigint, a: bigint, B: bigint): bigint {
         const N = this.parameters.primeGroup.N;
@@ -210,11 +210,11 @@ export class Routines {
     }
 
     /**
-     * Computes the session key S for the server.
+     * Computes the session key "S" for the server.
      * @param verifier The verifier.
      * @param u The U.
-     * @param A The client's public value.
-     * @param b The server's private value.
+     * @param A The client's public key "A".
+     * @param b The server's private key "b".
      */
     computeServerSessionKey(verifier: bigint, u: bigint, A: bigint, b: bigint): bigint {
         const N = this.parameters.primeGroup.N
