+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "@project-christopher/srp",
-  "version": "2.2.1",
+  "version": "2.2.2",
   "description": "",
   "main": "dist/index.js",
   "types": "dist/index.d.ts",
